Make CR in quickfix open a split instead of error
The quickfix often gives me errors because it's trying to jump to a file
by re-using a window with a modified file. I want it to go to the other
window and always seem to organize my windows wrong.

This makes it open a split instead of giving me an error.

Also minor cleanup: s: instead of SID and fix order of mapkey.