Don't resize quickfix if vertically split
Fix cmdheight becomes massive when quickfix is in a vertical split.

Using :resize on a window that uses the vim window's full height causes
the editable area to shrink to the new size and the cmd area fills in
the remaining space. Not sure why you'd ever want vim to have that behaviour.

Now we return whether we resized the quickfix.